- podmieniono temporaryWord na dictrionaryWord i usunieto zbedne pliki - dodane prototypowy widok slownictwa, oraz mozliwosc dodawania i usuwnaia slowa z fiszek - dodano widok zarzadzania slowami i grupami slow na administratora - dodano modale dodawania edycji itd.

+++ Arapp/src/main/java/engineer/arabski/wordBank/dto/WordGroupPreviewResponse.java
@@ -8,7 +8,7 @@ public record WordGroupPreviewResponse(
         String icon,
         String imageUrl,
         String category,
-        Long wordsCount
+        int wordsCount
 
 ) {
 }
